Separate TextureFactory from MaterialFactory
