+++ Assets/UniGLTF/Runtime/UniGLTF/MaterialFacotry/TextureLoader/GetTextureParam.cs
@@ -51,5 +51,4 @@ namespace UniGLTF
         }
     }
 
-    public delegate Task<Texture2D> GetTextureAsyncFunc(GetTextureParam param);
 }
