+++ Assets/UniGLTF/Runtime/UniGLTF/MaterialFacotry/TextureLoader/GetTextureParam.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 405597f56d6540347bbecb1203aba033
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
