+++ Assets/UniGLTF/Runtime/UniGLTF/MaterialFacotry/TextureLoader/TextureFactory.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8f1dc05e4970e724e9f50c0a5a67a25e
+guid: 4d6069f2bfbfc0c4c9276e8a5f8b0789
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
