+++ Assets/UniGLTF/Runtime/UniGLTF/TextureLoader.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a4bd2e8f388fb204186d743f337ddb83
+guid: 54a405d31cce94643be8f76f805da8c8
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
